addresses review comments
